requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
langchain
langchain-openai 
langgraph
langchain_community
pypdf
faiss-cpu

langchain-core 
langchain-text-splitters 
pinecone-client 
langchain-pinecone

boto3
uuid6

langgraph-checkpoint-aws

python-jose[cryptography]
requests

cryptography 
